chore(Refactor): Reorganize RulesTable component
